refactor: make designer flexible for other nodes
In this commit, we refactor the designer components to make it
possible for us to introduce another bitcoin node implementation
later.
