+++ src/components/designer/Sidebar.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
 import { IChart } from '@mrblenny/react-flow-chart';
-import { LightningNode, TapNode } from 'shared/types';
+import { BitcoindNode, LightningNode, TapNode } from 'shared/types';
 import { Network } from 'types';
-import BitcoindDetails from './bitcoind/BitcoindDetails';
+import BitcoindDetails from './bitcoin/BitcoinDetails';
 import DefaultSidebar from './default/DefaultSidebar';
 import LightningDetails from './lightning/LightningDetails';
 import LinkDetails from './link/LinkDetails';
@@ -21,7 +21,7 @@ const Sidebar: React.FC<Props> = ({ network, chart }) => {
       const { bitcoin, lightning, tap } = network.nodes;
       const node = [...bitcoin, ...lightning, ...tap].find(n => n.name === id);
       if (node && node.implementation === 'bitcoind') {
-        return <BitcoindDetails node={node} />;
+        return <BitcoindDetails node={node as BitcoindNode} />;
       } else if (node && node.type === 'lightning') {
         return <LightningDetails node={node as LightningNode} />;
       } else if (node && node.type === 'tap') {
