+++ src/components/designer/bitcoin/BitcoinDetails.tsx
