+++ src/components/designer/bitcoin/InfoTab.tsx
