+++ src/components/designer/bitcoin/actions/MineBlocksInput.spec.tsx
