+++ src/components/designer/bitcoin/actions/MineBlocksInput.tsx
